Wrote tests for login
